+++ Libplanet.Extensions.Cocona/Commands/ApvCommand.cs
@@ -1,18 +1,21 @@
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
-using System.Linq;
-using Bencodex;
-using Bencodex.Types;
+#pragma warning disable SA1200
 using Cocona;
-using Libplanet.Crypto;
-using Libplanet.KeyStore;
-using Libplanet.Net;
+#pragma warning restore SA1200
 
-namespace Libplanet.Tools
+namespace Libplanet.Extensions.Cocona.Commands
 {
-    public class Apv
+    using System;
+    using System.Collections.Generic;
+    using System.Globalization;
+    using System.IO;
+    using System.Linq;
+    using Bencodex;
+    using Bencodex.Types;
+    using Libplanet.Crypto;
+    using Libplanet.KeyStore;
+    using Libplanet.Net;
+
+    public class ApvCommand
     {
         [Command(Description = "Sign a new app protocol version.")]
         public void Sign(
@@ -45,7 +48,7 @@ namespace Libplanet.Tools
             string[]? extra = null
         )
         {
-            PrivateKey key = new Key().UnprotectKey(keyId, passphrase);
+            PrivateKey key = new KeyCommand().UnprotectKey(keyId, passphrase);
             IValue? extraValue = null;
             if (extraFile is string path)
             {
@@ -159,7 +162,7 @@ namespace Libplanet.Tools
 
             if (!noKeyStore)
             {
-                Key keyInstance = new Key();
+                KeyCommand keyInstance = new KeyCommand();
                 IEnumerable<Tuple<Guid, ProtectedPrivateKey>> ppks = keyInstance.KeyStore.List()
                     .Where(pair => pair.Item2.Address.Equals(v.Signer));
                 foreach (Tuple<Guid, ProtectedPrivateKey> pair in ppks)
