+++ Libplanet.Extensions.Cocona/Commands/KeyCommand.cs
@@ -1,16 +1,19 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
+#pragma warning disable SA1200
 using Cocona;
-using Libplanet.Crypto;
-using Libplanet.KeyStore;
+#pragma warning restore SA1200
 
-namespace Libplanet.Tools
+namespace Libplanet.Extensions.Cocona.Commands
 {
-    public class Key
+    using System;
+    using System.Collections.Generic;
+    using System.IO;
+    using System.Linq;
+    using Libplanet.Crypto;
+    using Libplanet.KeyStore;
+
+    public class KeyCommand
     {
-        public Key()
+        public KeyCommand()
         {
             KeyStore = Web3KeyStore.DefaultKeyStore;
         }
