+++ Libplanet.Extensions.Cocona/Libplanet.Extensions.Cocona.csproj
@@ -1,28 +1,18 @@
 <Project Sdk="Microsoft.NET.Sdk">
+
   <PropertyGroup>
-    <PackageId>Libplanet.Tools</PackageId>
-    <Title>planet: Libplanet CLI Tools</Title>
-    <Summary>planet: Libplanet CLI Tools</Summary>
-    <Description>This CLI app is a collection of utilities for application <!--
-    --> programmers who create decentralized games powered by Libplanet <!--
-    --> (https://libplanet.io/).</Description>
-    <OutputType>Exe</OutputType>
-    <PackageProjectUrl>https://libplanet.io/</PackageProjectUrl>
-    <PackageReadmeFile>README.md</PackageReadmeFile>
-    <PackageIcon>icon.png</PackageIcon>
+    <PackageId>Libplanet.Extensions.Cocona</PackageId>
     <Authors>Planetarium</Authors>
     <Company>Planetarium</Company>
     <PackageLicenseExpression>LGPL-2.1-or-later</PackageLicenseExpression>
     <RequireLicenseAcceptance>true</RequireLicenseAcceptance>
     <RepositoryUrl>git://github.com/planetarium/libplanet.git</RepositoryUrl>
-    <RepositoryType>git</RepositoryType>
   </PropertyGroup>
 
   <PropertyGroup>
     <LangVersion>8.0</LangVersion>
     <TargetFramework>netcoreapp3.1</TargetFramework>
-    <RootNamespace>Libplanet.Tools</RootNamespace>
-    <ToolCommandName>planet</ToolCommandName>
+    <RootNamespace>Libplanet.Extensions.Cocona</RootNamespace>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Nullable>enable</Nullable>
     <CheckForOverflowUnderflow>true</CheckForOverflowUnderflow>
@@ -31,29 +21,6 @@
     <IsTestProject>false</IsTestProject>
     <CodeAnalysisRuleSet>..\Libplanet.ruleset</CodeAnalysisRuleSet>
   </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(PublishSingleFile)' != 'true' ">
-    <PackAsTool>true</PackAsTool>
-    <AssemblyName>Libplanet.Tools</AssemblyName>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(PublishSingleFile)' == 'true' ">
-    <AssemblyName>planet</AssemblyName>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(PublishSingleFile)' == 'true' And&#xD;&#xA;                             '$(RuntimeIdentifier.Substring(0, 6))' == 'linux-' ">
-    <InvariantGlobalization>true</InvariantGlobalization>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Include="..\LICENSE" Pack="true" PackagePath="LICENSE.txt" />
-    <None Include="README.md" Pack="true" PackagePath="README.md" />
-    <None Include="..\icon.png" Pack="true" PackagePath="icon.png" />
-    <AdditionalFiles Include="..\Menees.Analyzers.Settings.xml">
-      <Link>Menees.Analyzers.Settings.xml</Link>
-    </AdditionalFiles>
-    <AdditionalFiles Include="..\stylecop.json" />
-  </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Cocona.Lite" Version="1.5.*" />
@@ -82,10 +49,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Libplanet.Extensions.Cocona\Libplanet.Extensions.Cocona.csproj" />
     <ProjectReference Include="..\Libplanet\Libplanet.csproj" />
     <ProjectReference Include="..\Libplanet.RocksDBStore\Libplanet.RocksDBStore.csproj" />
-    <!-- FIXME: We should specify the version range when the following NuGet
-    issue is addressed: <https://github.com/NuGet/Home/issues/5556>. -->
   </ItemGroup>
+
 </Project>
