+++ Libplanet.Tools/Program.cs
@@ -2,12 +2,14 @@ using System;
 using System.IO;
 using System.Threading.Tasks;
 using Cocona;
+using Libplanet.Extensions.Cocona.Commands;
+using Libplanet.Extensions.Cocona.Extensions;
 
 namespace Libplanet.Tools
 {
-    [HasSubCommands(typeof(Apv), Description = "App protocol version utilities.")]
-    [HasSubCommands(typeof(Key), Description = "Manage private keys.")]
-    [HasSubCommands(typeof(Mpt), Description = "Merkle Patricia Trie utilities.")]
+    [HasSubCommands(typeof(ApvCommand), "apv", Description = "App protocol version utilities.")]
+    [HasSubCommands(typeof(KeyCommand), "key", Description = "Manage private keys.")]
+    [HasSubCommands(typeof(MptCommand), "mpt", Description = "Merkle Patricia Trie utilities.")]
     public class Program
     {
         private static readonly string FileConfigurationServiceRoot = Path.Combine(
