+++ Libplanet.sln
@@ -23,6 +23,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Libplanet.Analyzers", "Libp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Libplanet.Analyzers.Tests", "Libplanet.Analyzers.Tests\Libplanet.Analyzers.Tests.csproj", "{B48782B9-6FA9-4D18-8564-1849FF4CF40E}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Libplanet.Extensions.Cocona", "Libplanet.Extensions.Cocona\Libplanet.Extensions.Cocona.csproj", "{B3170309-55AB-462C-9100-D77107799E82}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -153,6 +155,18 @@ Global
 		{B48782B9-6FA9-4D18-8564-1849FF4CF40E}.Release|x64.Build.0 = Release|Any CPU
 		{B48782B9-6FA9-4D18-8564-1849FF4CF40E}.Release|x86.ActiveCfg = Release|Any CPU
 		{B48782B9-6FA9-4D18-8564-1849FF4CF40E}.Release|x86.Build.0 = Release|Any CPU
+		{B3170309-55AB-462C-9100-D77107799E82}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B3170309-55AB-462C-9100-D77107799E82}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B3170309-55AB-462C-9100-D77107799E82}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B3170309-55AB-462C-9100-D77107799E82}.Debug|x64.Build.0 = Debug|Any CPU
+		{B3170309-55AB-462C-9100-D77107799E82}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B3170309-55AB-462C-9100-D77107799E82}.Debug|x86.Build.0 = Debug|Any CPU
+		{B3170309-55AB-462C-9100-D77107799E82}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B3170309-55AB-462C-9100-D77107799E82}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B3170309-55AB-462C-9100-D77107799E82}.Release|x64.ActiveCfg = Release|Any CPU
+		{B3170309-55AB-462C-9100-D77107799E82}.Release|x64.Build.0 = Release|Any CPU
+		{B3170309-55AB-462C-9100-D77107799E82}.Release|x86.ActiveCfg = Release|Any CPU
+		{B3170309-55AB-462C-9100-D77107799E82}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
